requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
base58==2.1.1
bitarray==1.2.2
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
cytoolz==0.11.2
dataclasses==0.8
eth-abi==2.1.1
eth-account==0.5.7
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
Flask==2.0.2
Flask-Cors==3.0.10
frozenlist==1.2.0
greenlet==1.1.2
hexbytes==0.2.2
idna==3.3
idna-ssl==1.1.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
ipfshttpclient==0.8.0a2
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==3.2.0
lru-dict==1.1.7
MarkupSafe==2.0.1
multiaddr==0.0.9
multidict==5.2.0
netaddr==0.8.0
parsimonious==0.8.1
protobuf==3.19.4
psycopg2==2.9.3
pycryptodome==3.14.1
pyrsistent==0.18.0
python-decouple==3.6
requests==2.27.1
rlp==2.0.1
six==1.16.0
SQLAlchemy==1.4.31
toolz==0.11.2
typing_extensions==4.0.1
urllib3==1.26.8
varint==1.0.2
web3==5.27.0
websockets==9.1
Werkzeug==2.0.2
yarl==1.7.2
zipp==3.6.0

pytest==7.1.0  # https://docs.pytest.org/en/stable/
black==22.1.0  # https://github.com/psf/black#
flake8==4.0.1  # https://github.com/PyCQA/flake8
flake8-isort==4.1.1  # https://github.com/gforcada/flake8-isort
pre-commit==2.15.0  # https://github.com/pre-commit/pre-commit
alembic==1.7.7  # https://alembic.sqlalchemy.org/en/latest/
